Queries

Documentation

addStudent :: Student -> Update StudentDb ()

Adds a student to the student's database, the primary key is the | students dni (identification document number).

addEmployee :: Employee -> Update EmployeeDb ()

Adds a employee to the employee's database, the primary key is the | employee dni (identification document number).

addDepartment :: Department -> Update DepartmentDb ()

Adds a department to the department's database, the primaty key is the | department DeptId.

addSchool :: School -> Update SchoolDb ()

Adds a schools to the schools database, the primary key is the | schools SchooId.

addResult :: Result -> Update ResultDb ()

Adds a result to the results database, increments the index.

addCompany :: Company -> Update CompanyDb ()

Adds a company to the companies database, increments the index.

addGrant :: Grant -> Update GrantDb ()

Adds a Grant to the grants database, increments the index.

addProject :: Project -> Update ProjectDb ()

Adds a Project to the project database, increments the index.

addNGO :: NGO -> Update NGODb ()

Adds a NGO to the ngos database, increments the index.

addGO :: GovernmentalOrganization -> Update GODb ()

Adds a GOD to the gods database, increments the index.

searchStudentDNI :: DNI -> Query StudentDb (Maybe Student)

Search for a student by the DNI.

searchEmployeeDNI :: DNI -> Query EmployeeDb (Maybe Employee)

Search for a employee by the DNI.

searchPubAuthor :: ByteString -> Query ResultDb [Result]

Search for a publication by Authors

searchPubKeyword :: ByteString -> Query ResultDb [Result]

Search for a publication by Keyword

getDeptAll :: Query DepartmentDb [(DeptId, Department)]

Gets all the departments.

getSchoolAll :: Query SchoolDb [(SchoolId, School)]

Gets all the schools.

getStudentAll :: Query StudentDb [(DNI, Student)]

Gets all the students.

getEmployeeAll :: Query EmployeeDb [(DNI, Employee)]

Gets all the Employees.

getResultAll :: Query ResultDb [Result]

Gets all Results.

getCompAll :: Query CompanyDb [Company]

Gets all the Companies.

getNGOAll :: Query NGODb [NGO]

Gets all the NGOs.

getGOAll :: Query GODb [GovernmentalOrganization]

Gets all the GOs.

getGrantAll :: Query GrantDb [Grant]

Gets all the Grants.

getProjectAll :: Query ProjectDb [Project]

Get all the Projects.

addStudentToDept :: DNI -> DeptId -> Update HasRelDb ()

Adds the relationship between a Student and a Department.

addDeptToSchool :: DeptId -> SchoolId -> Update HasRelDb ()

Adds the relationship between a School and a Department.

addSchoolToUni :: SchoolId -> UniId -> Update HasRelDb ()

Adds the relationship between a University and a School.

addProjectToUni :: ProjectId -> UniId -> Update HasRelDb ()

Adds the relationship between a University and a Project.

addGrantToUni :: GrantId -> UniId -> Update HasRelDb ()

Adds the relationship between a University and a Grant.

addGrantToProject :: GrantId -> ProjectId -> Update HasRelDb ()

Adds the relationship between a Project and a Grant.

addStudentToProject :: DNI -> ProjectId -> Update WorksOnDb ()

Adds the relationship between a Student and a Project

addEmployeeToProject :: DNI -> ProjectId -> Update WorksOnDb ()

Adds the relationship between a Employee and a Project

addCompanyToGrant :: CompanyId -> GrantId -> Update FundsDb ()

Adds the relationship between a Company and a Grant.

addNGOToGrant :: NGOId -> GrantId -> Update FundsDb ()

Adds the relationship between a NGO and A Grant.

addEmployeeToDept :: DNI -> DeptId -> Update EmploysDb ()

Adds an employee to a Department.

addEmployeeToSchool :: DNI -> SchoolId -> Update EmploysDb ()

Adds an employee to a School.

addEmployeeToUni :: DNI -> UniId -> Update EmploysDb ()

Adds an employee to the University.

studentsOnDept :: DeptId -> Query HasRelDb [DNI]

Looks for the students (DNI) that belongs to a given Department.

deptsOnSchool :: SchoolId -> Query HasRelDb [DeptId]

Looks for the departments that belongs to a given School.

schoolsOnUniversity :: UniId -> Query HasRelDb [SchoolId]

Looks for the schools inside a University.

projectsOnUniversity :: UniId -> Query HasRelDb [ProjectId]

Looks for the Projects inside the University.

grantsOnUniversity :: UniId -> Query HasRelDb [GrantId]

Looks for the grants inside a University.

grantsOnProject :: ProjectId -> Query HasRelDb [GrantId]

Looks for the grants associated to a Project.

projectsOnGrant :: GrantId -> Query HasRelDb [ProjectId]

Looks projects associated to a Grant.

studentsOnProject :: ProjectId -> Query WorksOnDb [DNI]

Looks for the Students working on a Project.

employeesOnProject :: ProjectId -> Query WorksOnDb [DNI]

Looks for the Employees working on a Project.

projectsOfStudent :: DNI -> Query WorksOnDb [ProjectId]

Looks for projects where a student works.

projectsOfEmployee :: DNI -> Query WorksOnDb [ProjectId]

Looks for the projects where an employee works.