Types

Documentation

newtype SchoolId

Constructors

SchoolId Integer

newtype DeptId

Constructors

DeptId Integer

type ProjectId = Integer

type GrantId = Integer

type ResultId = Integer

type GOId = Integer

type NGOId = Integer

type CompanyId = Integer

type UniId = Integer

newtype PatentId

Constructors

PatentId Integer

newtype PubId

Constructors

PubId Integer

newtype DNI

Constructors

DNI Integer

data Degree

Constructors

Bachelor

Master

PhD

PostPhD

data UniType

Constructors

Public

Private

data NType

Constructors

National

International

data Student

Constructors

Student

Fields

stdFname :: !ByteString

stdLname :: !ByteString

stdDni :: !DNI

stdPi :: !Bool

stdBirthday :: !Day

data Employee

Constructors

Employee

Fields

empFname :: !ByteString

empLname :: !ByteString

empDni :: !DNI

empPi :: !Bool

empBirthday :: !Day

empStudies :: !Degree

data Department

Constructors

Department

Fields

deptId :: !DeptId

deptName :: !ByteString

data School

Constructors

School

Fields

schoolId :: !SchoolId

schoolName :: !ByteString

students :: ![DNI]

employees :: ![DNI]

departments :: ![DeptId]

data NGO

Constructors

NGO

Fields

ngoName :: !ByteString

ngoType :: !NType

data GovernmentalOrganization

Constructors

GO

Fields

goId :: !GOId

goName :: !ByteString

data Grant

Constructors

Grant

Fields

grantId :: !GrantId

grantName :: !ByteString

grantLegislation :: !ByteString

grantTotalAmount :: !Double

data Project

Constructors

Project

Fields

projectId :: !ProjectId

projectName :: !ByteString

projectBudget :: !Double

projectGrants :: !GrantId

projectDesp :: !ByteString

data Result

Constructors

Publication

Fields

pubId :: !PubId

pubTitle :: !ByteString

pubDate :: !Day

pubAbstract :: !ByteString

pubJournal :: !ByteString

pubKeywords :: ![ByteString]

pubAuthors :: ![ByteString]

Patent

Fields

patentId :: !PatentId

patentName :: !ByteString

patentType :: !ByteString

patentNumber :: !ByteString

patentTerm :: !Integer

patentDate :: !Day

patentAuthors :: ![ByteString]

BiologicalOrganism

Molecule

Software

Hardware

data Company

Constructors

Company

Fields

companyIdimport, companyName :: !ByteString

companyType :: !NType

companyProfit :: !Double

data StudentDb

DATABASES----

Constructors

StudentDb (Map DNI Student)

data EmployeeDb

Constructors

EmployeeDb (Map DNI Employee)

data DepartmentDb

Constructors

DepartmentDb (Map DeptId Department)

data SchoolDb

Constructors

SchoolDb (Map SchoolId School)

data ResultDb

Constructors

ResultDb (Map ResultId Result)

data CompanyDb

Constructors

CompanyDb (Map CompanyId Company)

data NGODb

Constructors

NGODb (Map NGOId NGO)

data GODb

Constructors

GODb (Map GOId GovernmentalOrganization)

data GrantDb

Constructors

GrantDb (Map GrantId Grant)

data ProjectDb

Constructors

ProjectDb (Map ProjectId Project)

data University

Constructors

University

Fields

studentDb :: StudentDb

employeeDb :: EmployeeDb

departmentDb :: DepartmentDb

schoolDb :: SchoolDb

resultDb :: ResultDb

projectDb :: ProjectDb

grantDb :: GrantDb

name :: !ByteString

uniType :: !UniType

data Employs

Relationships

Defines the employer and employee relationship between a university, | department and school, and a employee. (Employer, Employee)

Constructors

EmployedByDept DeptId DNI

EmployedBySchool SchoolId DNI

EmployedByUniversity UniId DNI

data HasRel

Defines the has relationship between entities.

Constructors

DeptHasStudent DeptId DNI

SchoolHasDept SchoolId DeptId

UniversityHasSchool UniId SchoolId

UniversityHasProject UniId ProjectId

UniversityHasGrant UniId GrantId

ProjectHasGrant ProjectId GrantId

data WorksOn

Defines the relation works on between employee and student, and project.

Constructors

StudentWorksOn DNI ProjectId

EmployeeWorksOn DNI ProjectId

data Funds

Defines the funds relationship between company, go and ngo, and grants. | (go ngo company, grant).

Constructors

CompanyFunds CompanyId GrantId

NGOFunds NGOId GrantId

GOFunds GOId GrantId

data HasRelDb

The database for the hasRel relationships.

Constructors

HasRelDb

Fields

allHasRels :: [HasRel]

data WorksOnDb

The database for the worksOn relationships.

Constructors

WorksOnDb

Fields

allWorksOn :: [WorksOn]

data FundsDb

The database for the Funds relationships.

Constructors

FundsDb

Fields

allFunds :: [Funds]

data EmploysDb

The database for the employs relationships.

Constructors

EmploysDb

Fields

allEmploys :: [Employs]